requirements.txt
# Core
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0

# LLM
openai>=1.60.0
anthropic>=0.42.0

# LangGraph
langgraph>=0.2.0
langchain-core>=0.3.0

# Azure
azure-search-documents>=11.6.0
azure-identity>=1.19.0

# Document processing
python-pptx>=1.0.0
python-docx>=1.1.0
PyPDF2>=3.0.0
openpyxl>=3.1.0
pytesseract>=0.3.10
pdf2image>=1.17.0

# Local vector search
numpy>=2.1.0

# Utilities
aiohttp>=3.11.0
redis>=5.2.0
rich>=13.9.0

# Backend API (M11)
python-multipart>=0.0.9
sse-starlette>=2.0.0
httpx>=0.28.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0

# Code quality
ruff>=0.9.0
mypy>=1.14.0
